benchmark
```
without braces
patch:   3199358.0 i/s
master:  1166684.9 i/s - 2.74x  slower

with braces
patch:   2222942.3 i/s
master:   485430.5 i/s - 4.58x  slower
```