@ahorek 2018-09-10
ping @jeremy ?

I think it's much simplier and more readable than the original regex version. Let me know if you have any concerns.